Разделяет код на модули, main.js делает точкой входа
